refactor(sim): templatize ExpressionCache
Will be used to cache both terms and types.
