+++ lib/sim/debug/expr_parser.hpp
@@ -2,36 +2,11 @@
 #include <QtCore>
 #include <set>
 #include "expr_ast.hpp"
+#include "expr_cache.hpp"
 
 namespace pepp::debug {
 
-struct ExpressionCache {
-  struct Compare {
-    using is_transparent = void;
-    bool operator()(const std::shared_ptr<Term> &lhs, const std::shared_ptr<Term> &rhs) const { return *lhs < *rhs; }
-    bool operator()(const Term &lhs, const std::shared_ptr<Term> &rhs) const { return lhs < *rhs; }
-    bool operator()(const std::shared_ptr<Term> &lhs, const Term &rhs) const { return *lhs < rhs; }
-    bool operator()(const Term &lhs, const Term &rhs) const { return lhs < rhs; }
-  };
-  using Set = std::set<std::shared_ptr<Term>, Compare>;
-  template <typename T> std::shared_ptr<T> add_or_return(T &&item) {
-    QMutexLocker locker(&_mut);
-    Set::iterator search = _set.find(item);
-    if (search == _set.end()) {
-      auto ret = std::make_shared<T>(std::forward<T>(item));
-      // Set up dependent tracking on creation.
-      ret->link();
-      _set.insert(ret);
-      return ret;
-    } else return std::dynamic_pointer_cast<T>(*search);
-  }
-  void collect_garbage();
-  std::size_t count() const;
-
-private:
-  mutable QMutex _mut;
-  Set _set{};
-};
+using ExpressionCache = Cache<Term>;
 
 namespace detail {
 struct TokenBuffer;
